refactor: API 관련 패키지 구조 개선
- response와 request 클래스를 위한 별도 패키지 생성
- controller 상위에 api 패키지 추가
