+++ src/main/java/com/hh99/ecommerce/balance/api/request/ChargeBalanceRequest.java
@@ -1,14 +1,16 @@
-package com.hh99.ecommerce.balance.controller;
+package com.hh99.ecommerce.balance.api.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ChargeBalanceRequest {
     private Long userId;
-    private int amount;
+    private BigDecimal amount;
 }
